backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
python-multipart
google-generativeai
cors
psutil
requests
